Some small clean ups in movement logic, botlogic. Movement logic does not use any of the old methods now (and should never again!) so some of the fixes michele did might not work and we might have to sit together to solve movement problems in the right locations.
A bigger new method in blackboard for priorizing ammunition search - should probably move somewhere else at some point, but lets see where this goes. 

ActorSystem now has a  tiny if-else thingy to test the new actions - of course they do not really make sense this way and should be decided upon by decision logic.

actions:

suppressionFire: this just shoots where bot is looking, using machinegun, assaultrifle or flak cannon in this order and in primary mode. It is meant for example during retreat, as distraction, or as fallback if other actions are not useable.
randomWalk: go to random position (PROBLEM: not navmesh but location!!! this might be an issue)
gotoammunition: go to the most important ammunition location that is found on the current map
followvisibleplayer: as the name suggests. jsut like before, trivial.